IT WORKS!
fixed many, many bugs, implemented playing (should be relatively
efficient), volume envelopes (but not panning envelopes). found that
instruments start at 1, not 0, and note C-0 is 1, not 0... so it is now
in tune with FT2!

no effects are supported at all yet, and there's a bug where it stops
playing entirely for a while, but "my dirty old kamel" sounds
recognizable and in tune!